superheroes/superhero-04-02/requirements.txt
flask
flask_cors
google.generativeai
python-dotenv
firebase-admin